+++ src/main/java/net/geforcemods/securitycraft/gui/components/IToggleableButton.java
@@ -1,0 +1,7 @@
+package net.geforcemods.securitycraft.gui.components;
+
+public interface IToggleableButton {
+	public int getCurrentIndex();
+
+	public void setCurrentIndex(int newIndex);
+}